+++ pot.cpp
@@ -3,7 +3,7 @@
 #include "setup.h"
 
 float  factor_F;
-int getDifficult(){
+byte getDifficult(){
   int  readedPot =  analogRead(POTPIN);
   return map(readedPot,0,1024,1,9);
 }
@@ -11,6 +11,4 @@ int getDifficult(){
 void calculateFactorF(){ // choice diffult: max diffucult = calculatedPercentage : 50(max percentage)
   float difficult = getDifficult();
   factor_F = (difficult * 50)/8;
-  
-  
 }
